+++ src/pwm/PwmController.cpp
@@ -1,7 +1,7 @@
 #include <Arduino.h>
 #include <pwm/PwmController.h>
 
-PwmController::PwmController(Pwm* pwm, const int redPin, const int greenPin, const int bluePin): pwm(pwm), redPin(redPin), greenPin(greenPin), bluePin(bluePin) {
+PwmController::PwmController(Pwm* pwm, const uint8_t redPin, const uint8_t greenPin, const uint8_t bluePin): pwm(pwm), redPin(redPin), greenPin(greenPin), bluePin(bluePin) {
     pinMode(redPin, OUTPUT);
     pinMode(greenPin, OUTPUT);
     pinMode(bluePin, OUTPUT);
